Revert "0.5.2"
This reverts commit f69ea0cbedc64b3d396ffd9057b7dd80521bcf8d.

+++ lib/opzworks/commands/berks.rb
@@ -68,20 +68,26 @@ module OpzWorks
           s3_bucket        = config.berks_s3_bucket || 'opzworks'
           overrides        = 'overrides'
 
-          FileUtils.mkdir_p("#{install_path}") unless File.directory?("#{install_path}")
-          FileUtils.copy("#{@target_path}/Berksfile.opsworks", "#{install_path}/Berksfile")
+          if File.exist?("#{@target_path}/Berksfile.opsworks")
+            puts 'Remote management berksfile detected, not building local berkshelf.'.foreground(:yellow)
 
-          # berks
-          #
-          puts 'Running berks install'.foreground(:blue)
-          run_local <<-BASH
-            cd #{@target_path}
-            berks update
-          BASH
-          run_local <<-BASH
-            cd #{@target_path}
-            berks vendor #{install_path}
-          BASH
+            unless File.directory?("#{install_path}")
+              FileUtils.mkdir_p("#{install_path}")
+            end
+            FileUtils.copy("#{@target_path}/Berksfile.opsworks", "#{install_path}/Berksfile")
+          else
+            # berks
+            #
+            puts 'Running berks install'.foreground(:blue)
+            run_local <<-BASH
+              cd #{@target_path}
+              berks update
+            BASH
+            run_local <<-BASH
+              cd #{@target_path}
+              berks vendor #{install_path}
+            BASH
+          end
 
           # if there's an overrides file, just pull it and stuff the contents into the
           #   upload repo; the line is assumed to be a git repo. This is done to override
